+++ embedded_config_annotations/CHANGELOG.md
@@ -1,5 +1,9 @@
+## v0.3.0
+- **(breaking change)** Migrated to null-safety.
+- Changed minimum SDK version to `2.12.0`.
+
 ## v0.2.0
-- **(Breaking change)** `EmbeddedConfig.path` is now a `List<String>`. This addresses incompatibilities with keys containing `.` characters.
+- **(breaking change)** `EmbeddedConfig.path` is now a `List<String>`. This addresses incompatibilities with keys containing `.` characters.
 
 ## v0.1.1
 - Add an example to address package scoring.